Refactor SentryHandler and fix allocated bytes type
Removed unused imports and simplified if statements in SentryHandler for better code readability. Changed AllocatedBytes type from int to long in SentryModulesDto to handle larger values.

+++ src/Gml.Web.Api.Dto/Sentry/SentryModulesDto.cs
@@ -124,7 +124,7 @@ public class Image
 
 public class MemoryInfo
 {
-    [JsonProperty("allocated_bytes")] public int AllocatedBytes { get; set; }
+    [JsonProperty("allocated_bytes")] public long AllocatedBytes { get; set; }
 
     [JsonProperty("high_memory_load_threshold_bytes")]
     public long HighMemoryLoadThresholdBytes { get; set; }
